Add support for editing elements with role="textbox"
This is somewhat a hack to make IAT to work with gmail's new compose.
It might not work in every version of every browser but at least it
works for my specific use case. Unfortunately I couldn't get the
gumdrop icon to appear in gmail but the hot-key still works.

Signed-off-by: Patrik Jakobsson <patrik.r.jakobsson@gmail.com>